helpdesk_full/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.4
langchain==0.2.12
langchain-openai==0.1.22
langchain-pinecone==0.1.3
pinecone-client==5.0.1
python-dotenv==1.0.1
tiktoken==0.7.0
